Initialize hours_worked before using it
After a few hours of unsuccessful attempts at locating the cause of the
valgrind error:

	Syscall param exit(error_code) contains uninitialised byte(s)

I finally figured out that the hours_worked variable was being used
uninitialized ... (duh).

Case in point,

	hours_worked += nvtab.nameval[i].hours;

which expands to,

	hours_worked = hours_worked + nvtab.nameval[i].hours;

naturally.  Of course, hours_worked contains garbage until it's
initialized, so trying to use it beforehand was idiotic, at best.

Signed-off-by: Ian Hilt <ihilt@mcgregor-surmount.com>